# openapi.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}}}},"/users":{"post":{"operationId":"UsersController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseActionReturn"}}}}},"tags":["users"]},"get":{"operationId":"UsersController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserResponse"}}}}}},"tags":["users"]},"patch":{"operationId":"UsersController_update","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseActionReturn"}}}}},"tags":["users"],"security":[{"bearer":[]}]},"delete":{"operationId":"UsersController_remove","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseActionReturn"}}}}},"tags":["users"],"security":[{"bearer":[]}]}},"/users/{id}":{"get":{"operationId":"UsersController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}}},"tags":["users"]}},"/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginResponseDto"}}}}},"tags":["auth"]}},"/auth/me":{"get":{"operationId":"AuthController_whoami","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginResponseDto"}}}}},"tags":["auth"],"security":[{"bearer":[]}]}},"/knowledge-bases":{"post":{"operationId":"KnowledgeBasesController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateKnowledgeBaseDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseActionReturn"}}}}},"tags":["knowledge-bases"],"security":[{"bearer":[]}]},"get":{"operationId":"KnowledgeBasesController_findAll","parameters":[{"name":"page","required":false,"in":"query","schema":{"type":"number"}},{"name":"limit","required":false,"in":"query","schema":{"type":"number"}},{"name":"title","required":false,"in":"query","schema":{"type":"string"}},{"name":"difficulty","required":false,"in":"query","schema":{"type":"string"}},{"name":"subject","required":false,"in":"query","schema":{"type":"string"}},{"name":"levelOfDetail","required":false,"in":"query","schema":{"type":"string"}},{"name":"userId","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/KnowledgeBaseResponse"}}}}}},"tags":["knowledge-bases"]}},"/knowledge-bases/{baseId}":{"patch":{"operationId":"KnowledgeBasesController_update","parameters":[{"name":"baseId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateKnowledgeBaseDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseActionReturn"}}}}},"tags":["knowledge-bases"],"security":[{"bearer":[]}]},"delete":{"operationId":"KnowledgeBasesController_remove","parameters":[{"name":"baseId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseActionReturn"}}}}},"tags":["knowledge-bases"],"security":[{"bearer":[]}]},"get":{"operationId":"KnowledgeBasesController_findOne","parameters":[{"name":"baseId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/KnowledgeBaseResponse"}}}}},"tags":["knowledge-bases"]}},"/notes/{baseId}":{"post":{"operationId":"NotesController_generate","parameters":[{"name":"baseId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FullGenerateNotesDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseActionReturn"}}}}},"tags":["notes"],"security":[{"bearer":[]}]}},"/notes":{"get":{"operationId":"NotesController_findAll","parameters":[{"name":"page","required":false,"in":"query","schema":{"type":"number"}},{"name":"limit","required":false,"in":"query","schema":{"type":"number"}},{"name":"title","required":false,"in":"query","schema":{"type":"string"}},{"name":"knowledgeBaseId","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/NoteResponse"}}}}}},"tags":["notes"]}},"/notes/{noteId}":{"get":{"operationId":"NotesController_findOne","parameters":[{"name":"noteId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NoteResponseExtended"}}}}},"tags":["notes"]}},"/notes/{baseId}/{noteId}":{"patch":{"operationId":"NotesController_update","parameters":[{"name":"noteId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateNoteDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseActionReturn"}}}}},"tags":["notes"],"security":[{"bearer":[]}]},"delete":{"operationId":"NotesController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseActionReturn"}}}}},"tags":["notes"],"security":[{"bearer":[]}]}},"/blob/images":{"post":{"operationId":"BlobController_uploadImage","parameters":[],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"files":{"type":"array","items":{"type":"string","format":"binary"}}}}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/FileResult"}}}}}},"tags":["blob"],"security":[{"bearer":[]}]}},"/blob/audio":{"post":{"operationId":"BlobController_uploadAudio","parameters":[],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"files":{"type":"array","items":{"type":"string","format":"binary"}}}}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/FileResult"}}}}}},"tags":["blob"],"security":[{"bearer":[]}]}},"/blob/video":{"post":{"operationId":"BlobController_uploadVideo","parameters":[],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"files":{"type":"array","items":{"type":"string","format":"binary"}}}}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/FileResult"}}}}}},"tags":["blob"],"security":[{"bearer":[]}]}},"/math-problems":{"post":{"operationId":"MathProblemsController_uploadImage","parameters":[],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"type":"object","properties":{"file":{"type":"string","format":"binary"}}}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseActionReturn"}}}}},"tags":["math-problems"],"security":[{"bearer":[]}]},"get":{"operationId":"MathProblemsController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MathProblemResponse"}}}}}},"tags":["math-problems"]}},"/math-problems/me":{"get":{"operationId":"MathProblemsController_findAllForUser","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MathProblemResponse"}}}}}},"tags":["math-problems"],"security":[{"bearer":[]}]}},"/math-problems/{id}":{"get":{"operationId":"MathProblemsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MathProblemResponse"}}}}},"tags":["math-problems"]},"delete":{"operationId":"MathProblemsController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseActionReturn"}}}}},"tags":["math-problems"]}},"/question-answers":{"post":{"operationId":"QuestionAnswersController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateQuestionAnswerDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BaseActionReturn"}}}}},"tags":["question-answers"],"security":[{"bearer":[]}]},"get":{"operationId":"QuestionAnswersController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/QuestionAnswerResponse"}}}}}},"tags":["question-answers"]}},"/question-answers/me":{"get":{"operationId":"QuestionAnswersController_findMe","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/QuestionAnswerResponse"}}}}}},"tags":["question-answers"],"security":[{"bearer":[]}]}},"/question-answers/{id}":{"get":{"operationId":"QuestionAnswersController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/QuestionAnswerResponse"}}}}},"tags":["question-answers"]}}},"info":{"title":"Power Study API","description":"Api for the power study application","version":"0.1","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"CreateUserDto":{"type":"object","properties":{"name":{"type":"string"},"password":{"type":"string"},"email":{"type":"string"},"profilePicture":{"type":"string"}},"required":["name","password","email"]},"BaseActionReturn":{"type":"object","properties":{"id":{"type":"number"}},"required":["id"]},"UserResponse":{"type":"object","properties":{"id":{"type":"number","readOnly":true},"email":{"type":"string","readOnly":true},"name":{"type":"string","readOnly":true},"profilePicture":{"type":"string"}},"required":["id","email","name"]},"UpdateUserDto":{"type":"object","properties":{"name":{"type":"string"},"password":{"type":"string"},"email":{"type":"string"},"profilePicture":{"type":"string"}}},"LoginDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"LoginResponseDto":{"type":"object","properties":{"accessToken":{"type":"string"}},"required":["accessToken"]},"CreateKnowledgeBaseDto":{"type":"object","properties":{"title":{"type":"string"},"difficulty":{"type":"string"},"levelOfDetail":{"type":"string"},"subject":{"type":"string"},"language":{"type":"string"}},"required":["title","difficulty","levelOfDetail","subject"]},"KnowledgeBaseResponse":{"type":"object","properties":{"id":{"type":"number","readOnly":true},"title":{"type":"string","readOnly":true},"difficulty":{"type":"string","readOnly":true},"levelOfDetail":{"type":"string","readOnly":true},"subject":{"type":"string","readOnly":true}},"required":["id","title","difficulty","levelOfDetail","subject"]},"UpdateKnowledgeBaseDto":{"type":"object","properties":{"title":{"type":"string"},"difficulty":{"type":"string"},"levelOfDetail":{"type":"string"},"subject":{"type":"string"},"language":{"type":"string"}}},"GenerateNotesDto":{"type":"object","properties":{"fileUri":{"type":"string"},"mimeType":{"type":"string"}},"required":["fileUri","mimeType"]},"FullGenerateNotesDto":{"type":"object","properties":{"notes":{"type":"array","items":{"$ref":"#/components/schemas/GenerateNotesDto"}}},"required":["notes"]},"NoteResponse":{"type":"object","properties":{"id":{"type":"number"},"title":{"type":"string"},"content":{"type":"string","readOnly":true},"amountOfBookmarks":{"type":"number"}},"required":["id","title","content"]},"NoteResponseExtended":{"type":"object","properties":{"id":{"type":"number"},"title":{"type":"string"},"content":{"type":"string","readOnly":true},"amountOfBookmarks":{"type":"number"},"userId":{"type":"number"},"name":{"type":"string"},"profilePicture":{"type":"string"},"updatedAt":{"format":"date-time","type":"string"},"knowledgeBaseId":{"type":"number"},"knowledgeBaseTitle":{"type":"string"}},"required":["id","title","content","userId","name","profilePicture","updatedAt","knowledgeBaseId","knowledgeBaseTitle"]},"UpdateNoteDto":{"type":"object","properties":{"title":{"type":"string"},"content":{"type":"string"}}},"FileResult":{"type":"object","properties":{"fileUri":{"type":"string"},"mimeType":{"type":"string"}},"required":["fileUri","mimeType"]},"MathProblemResponse":{"type":"object","properties":{"id":{"type":"number","readOnly":true},"mathQuestion":{"type":"string","readOnly":true},"solution":{"type":"string","readOnly":true},"updatedAt":{"format":"date-time","type":"string","readOnly":true}},"required":["id","mathQuestion","solution","updatedAt"]},"CreateQuestionAnswerDto":{"type":"object","properties":{"question":{"type":"string"}},"required":["question"]},"QuestionAnswerResponse":{"type":"object","properties":{"id":{"type":"number"},"question":{"type":"string"},"answer":{"type":"string"},"updatedAt":{"format":"date-time","type":"string"}},"required":["id","question","answer","updatedAt"]}}}}